third_party/WebKit/LayoutTests/http/tests/serviceworker/fetch-body-stream.html

<!DOCTYPE html>
<title>Service Worker: fetch()</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js"></script>
<script>
var test = async_test('Verify FetchBodyStream in a Service Worker');

test.step(function() {
    var scope = 'resources/blank.html';
    service_worker_unregister_and_register(
        test, 'resources/fetch-body-stream-worker.js', scope).then(test.step_func(onRegister));

    function onRegister(worker) {
        var messageChannel = new MessageChannel();
        messageChannel.port1.onmessage = test.step_func(onMessage);
        worker.postMessage({port: messageChannel.port2}, [messageChannel.port2]);
    }

    var result = [];
    var expected = [
        'Text: <!DOCTYPE html>\n',
        'JSON: 1 2',
        'JSON: SyntaxError',
        'Blob: 16 : text/html',
        'ArrayBuffer: 16',
        '<!DOCTYPE html>\n : TypeError',
    ];

    function onMessage(e) {
        var message = e.data;
        if (message == 'quit') {
            assert_array_equals(result, expected, 'Worker should post back expected values.');
            service_worker_unregister_and_done(test, scope);
        } else {
            result.push(message);
        }
    }
});
</script>
